rename to protected and rename Slicer class to DimensionSlicer to reduce confusion
