+++ src/plopp/plotting/_inspector.py
@@ -13,8 +13,8 @@ from ..core.typing import Plottable
 from ..core.utils import coord_as_bin_edges
 from ..graphics import imagefigure, linefigure
 from ..widgets import Box, PointsTool, PolygonTool, RectangleTool
+from ._slicer import SlicerPlot
 from .common import preprocess, require_interactive_figure
-from .slicer_plot import SlicerPlot
 
 
 def _to_bin_edges(da: sc.DataArray, dim: str) -> sc.DataArray:
