+++ src/plopp/plotting/_scatter.py
